WIP - parse specific field-value object prop types
Still need to figure out how to model prop types as a potentially nesting structure of types. Currently types are bound up in Prop objects, which are a little too heavy to nest inside each other. I think the next step will be to tease apart the concepts of "prop" and "prop type".