Add support for array-type props (not working!)
Currently this only works for primitive array types like number[] or
string[]. References to custom types (e.g. Analytics[]) do not get the
type definition expanded by the Stencil compiler, which prevents us from
parsing the type fully.

Need to dig through the internals of Stencil a bit more to see if this
will be possible to support.