Fix error running Jest after npm build
The production build was including Jest tests in the compiled output,
which was causing Jest to run those compiled tests along with the
originals in the source directory, and failing due to the ES imports
in the compiled versions.

This creates a special TypeScript config file for npm run build that
excludes the tests, and enables the "main" TypeScript config to produce
the Jest-compatible CommonJS modules.